+++ deltakit-core/pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
   "networkx>=3.1, <4.0",
   "stim>=1.12.1, <2.0",
   "typing_extensions>=4.0, <5.0",
+  "tqdm >=4.66.2",
 ]
 
 [project.urls]
