+++ deltakit-core/tests/unit/api/test_decorators.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-from deltakit_explorer._utils._decorators import validate_and_split_decoding, validate_generation
+from deltakit_core.api._decorators import validate_and_split_decoding, validate_generation
 from deltakit_core.types._types import DetectionEvents, ObservableFlips, Decoder
 from deltakit_core.api.enums import DecoderType, QECECodeType
 from deltakit_core.types._experiment_types import QECExperimentDefinition, CircuitParameters
