+++ deltakit-explorer/pyproject.toml
@@ -36,7 +36,6 @@ dependencies = [
     "matplotlib >=3.5.2",
     "requests >=2.32.0",
     "seaborn >=0.12.0",
-    "tqdm >=4.66.2",
     "typing-extensions >=4.0.0",
     "requests_toolbelt >=1.0.0",  # needed by gql, which doesn't automatically install it
     "pymatching >= 2.2.2",
