+++ deltakit-explorer/src/deltakit_explorer/_api/_client.py
@@ -17,7 +17,7 @@ from deltakit_explorer._api._api_v2_client import APIv2Client
 from deltakit_explorer._api._auth import TOKEN_VARIABLE
 from deltakit_explorer._api._gql_client import GQLClient
 from deltakit_core.api import environment as utils
-from deltakit_explorer._utils._decorators import (
+from deltakit_core.api._decorators import (
     validate_and_split_decoding, validate_and_split_simulation,
     validate_generation)
 from deltakit_core.api._logging import Logging
